add token authentication

+++ accountservice/src/main/java/com/example/accountservice/service/AccountService.java
@@ -1,12 +1,20 @@
 package com.example.accountservice.service;
 
 import com.example.accountservice.exception.AccountNotFoundException;
+import com.example.accountservice.exception.EmailExistsException;
 import com.example.accountservice.model.Account;
 import com.example.accountservice.repository.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpStatusCodeException;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -17,22 +25,20 @@ public class AccountService {
     private AccountRepository repo;
 
 
-    private Account acc;
-
     public List<Account> getAllAccount() {
         return repo.findAll();
     }
 
-    public Account getAccountById(int id) {
-        return repo.findById(id).orElseThrow();
+    public Optional<Account> getAccountById(int id) {
+        return repo.findById(id);
     }
 
-    public Account addAccount(Account account) {
-        return repo.save(account);
+    public Optional<Account> getAccountByName(String name) {
+        return Optional.ofNullable(repo.findByName(name));
     }
 
     public Account updateAccountById(int id, Account account) {
-        Account updateAccount = repo.findById(id).orElseThrow();
+        Account updateAccount = repo.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST));
 
         updateAccount.setName(account.getName());
         updateAccount.setSurname(account.getSurname());
@@ -42,31 +48,18 @@ public class AccountService {
         updateAccount.setUsername(account.getUsername());
         updateAccount.setEmail(account.getEmail());
         updateAccount.setPassword(account.getPassword());
+        updateAccount.setRole(account.getRole());
 
         return  repo.save(updateAccount);
     }
 
-    public String deleteAccountById(int id) {
-        repo.deleteById(id);
-        return "Account deleted";
-    }
-
-    //check if account is present on db
-    public String checkAccount(Account account) {
-
-        try {
-            // find the user and password on db
-            Account username = repo.findByUsername(account.getUsername());
-            Account password = repo.findByPassword(account.getPassword());
-
-            // this check if user and pass is null and launch the exception
-            username.getUsername();
-            password.getPassword();
-        } catch(Exception e) {
-            throw new AccountNotFoundException("Account doesn't exist");
+    public void deleteAccountById(int id) {
+        if (!repo.existsById(id)) {
+            throw new NoSuchElementException("Account with ID " + id + " not found");
         }
-
-        return "Login succes";
-
+        repo.deleteById(id);
     }
+
+
+
 }
